+++ src/App.js
@@ -1,28 +1,14 @@
 import * as React from 'react';
-
-const todoList =  [
-  {
-    id: 1,
-    title: 'Complete Chemistry Homework',
-  },
-  {
-    id: 2,
-    title: 'Respond to unread emails'
-  },
-  {
-    id: 3,
-    title: 'Sign up for a coding bootcamp'
-  },
-];
+import TodoList from './TodoList';
+import AddTodoForm from './AddTodoForm';
 
 function App() {
   
   return (
     <div>
       <h1>Todo List</h1>
-      <ul>
-            {todoList.map((item) => <li key={item.id}> {item.title} </li>)}
-        </ul>  
+      <AddTodoForm />
+      <TodoList /> 
     </div> 
   );
 }
